python_microservice/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
supabase==2.3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
aiofiles==23.2.1